+++ Tests/LinuxMain.swift
@@ -1,0 +1,18 @@
+// Generated using Sourcery 0.5.3 — https://github.com/krzysztofzablocki/Sourcery
+// DO NOT EDIT
+
+
+import XCTest
+import PromptLineTests
+
+
+extension PromptLineTapsTests {
+  static var allTests = [
+    ("testSomething", testSomething),
+  ]
+}
+
+
+XCTMain([
+  testCase(PromptLineTapsTests.allTests),
+])