+++ Tests/PromptLineTests/PromptLineTests.swift
@@ -1,0 +1,18 @@
+import XCTest
+import Taps
+import PromptLine
+
+public func describePromptLine() -> Never {
+  Taps.runMain(testing: [
+    describePrefixPromptRunner,
+    describeInfixPromptRunner,
+    describeFlatMap,
+    describeBashShell
+  ])
+}
+
+public class PromptLineTapsTests: XCTestCase {
+  public func testSomething() {
+    describePromptLine()
+  }
+}