+++ CHANGELOG.md
@@ -1,4 +1,22 @@
-# Prompt
+# PromptLine
+
+## 0.5.0
+
+### Breaking Changes
+
+- Reintroduced `TerminationReason` for `PromptError`
+- Renamed infix `%` to infix `>-`
+
+### API Additions
+
+- Intoduces `Prompt.Shell` and `Prompt.defaultShell`, which is great for Unit Tests
+
+### Further Changes
+
+- Uses Taps for Unit Tests
+- `>-(_:)` for `String` now executes in its own bash
+- All operators are now lazy when creating closures
+- Supports Linux by always using `.exit` as termination reason
 
 ## 0.4.0